fix: resolve records by type-qualified reference and fail loud on cross-type collision
The shared record resolver mapped a bare NNNN to the first matching
doc-type directory, so status, describe, and supersede silently mutated
the wrong record when a number collided across directories (describe 0029
overwrote ADR 0029; status 0028 closed targeted ADR 0028). Add a
type-qualified reference form TYPE/NNNN (reusing the registry token via
paths::dir_for); an unqualified cross-type collision now exits non-zero
listing every candidate and writes nothing. Single-directory resolution
is unchanged.

supersede.rs and status.rs crossed the 300-line file-size ratchet after
the new tests, so their inline test modules move to sibling tests.rs
files (mod.rs + tests.rs) per CLAUDE.md rule 5, with no test semantics
changed. Closes living-docs issues 0029 and 0025.
