+++ living-docs-core/src/commands/describe.rs
@@ -1,11 +1,11 @@
-use crate::commands::supersede::{find_record, parse_record_number, set_frontmatter_fields};
+use crate::commands::supersede::{find_record, set_frontmatter_fields};
 use crate::record::format_scalar;
 use crate::store::DocStore;
 use std::path::Path;
 use std::process::ExitCode;
 
-pub fn run(store: &dyn DocStore, docs_dir: &Path, number: &str, description: &str) -> ExitCode {
-    match describe(store, docs_dir, number, description) {
+pub fn run(store: &dyn DocStore, docs_dir: &Path, reference: &str, description: &str) -> ExitCode {
+    match describe(store, docs_dir, reference, description) {
         Ok(()) => ExitCode::SUCCESS,
         Err(message) => {
             eprintln!("living-docs describe: {message}");
@@ -15,23 +15,23 @@ pub fn run(store: &dyn DocStore, docs_dir: &Path, number: &str, description: &st
 }
 
 /// Sets a record's `description:` frontmatter field, reusing `supersede`'s
-/// record-resolution ([`find_record`]) and frontmatter-mutation
-/// ([`set_frontmatter_fields`]) helpers rather than duplicating them (lesson
-/// 3717), exactly as `status.rs` does. Unlike `status`, no vocabulary
-/// constrains `description` — any sentence is accepted (issue 0021, part 2
-/// of 2) — so the only failure mode is an unresolvable `number`, surfaced by
-/// `find_record` before any write is attempted. `description` is quoted via
-/// [`format_scalar`] exactly as `new.rs`'s `fill_frontmatter_description`
-/// quotes it, so a colon-bearing sentence stays a canonical-check fixed
-/// point.
+/// collision-aware record-resolution ([`find_record`], accepting either a
+/// bare `NNNN` or a type-qualified `TYPE/NNNN` reference — issue 0029/0025)
+/// and frontmatter-mutation ([`set_frontmatter_fields`]) helpers rather than
+/// duplicating them (lesson 3717), exactly as `status.rs` does. Unlike
+/// `status`, no vocabulary constrains `description` — any sentence is
+/// accepted (issue 0021, part 2 of 2) — so the only failure mode is an
+/// unresolvable `reference`, surfaced by `find_record` before any write is
+/// attempted. `description` is quoted via [`format_scalar`] exactly as
+/// `new.rs`'s `fill_frontmatter_description` quotes it, so a colon-bearing
+/// sentence stays a canonical-check fixed point.
 fn describe(
     store: &dyn DocStore,
     docs_dir: &Path,
-    number: &str,
+    reference: &str,
     description: &str,
 ) -> Result<(), String> {
-    let record_number = parse_record_number(number)?;
-    let path = find_record(store, docs_dir, record_number)?;
+    let path = find_record(store, docs_dir, reference)?;
     set_frontmatter_fields(store, &path, &[("description", format_scalar(description))])
 }
 
@@ -158,6 +158,78 @@ mod tests {
             .read(Path::new("/bundle/adr/0001-a-decision.md"))
             .unwrap();
         assert_eq!(unchanged, RECORD);
+    }
+
+    const ADR_RECORD_0029: &str =
+        "---\ntype: ADR\ntitle: A Decision\ndescription: <One sentence — the decision and its scope.>\nstatus: Proposed\nsupersedes:\nsuperseded_by:\n---\n\n# A Decision\n";
+    const ISSUE_RECORD_0029: &str =
+        "---\ntype: Issue\ntitle: A Bug\ndescription: <One sentence.>\nstatus: open\n---\n\n# A Bug\n";
+
+    #[test]
+    fn describe_resolves_a_type_qualified_reference_to_only_that_type_across_a_number_collision() {
+        let store = MapStore::seeded(&[
+            ("/bundle/adr/0029-collision.md", ADR_RECORD_0029),
+            ("/bundle/issues/0029-collision.md", ISSUE_RECORD_0029),
+        ]);
+
+        describe(
+            &store,
+            Path::new("/bundle"),
+            "adr/0029",
+            "A concise description.",
+        )
+        .expect("a type-qualified reference must resolve");
+
+        let adr = store
+            .read(Path::new("/bundle/adr/0029-collision.md"))
+            .unwrap();
+        let issue = store
+            .read(Path::new("/bundle/issues/0029-collision.md"))
+            .unwrap();
+        assert!(
+            adr.contains("description: A concise description.\n"),
+            "got: {adr}"
+        );
+        assert_eq!(
+            issue, ISSUE_RECORD_0029,
+            "the colliding Issue must stay byte-identical"
+        );
+    }
+
+    #[test]
+    fn describe_fails_loud_on_an_unqualified_cross_type_collision_and_writes_no_file() {
+        let store = MapStore::seeded(&[
+            ("/bundle/adr/0029-collision.md", ADR_RECORD_0029),
+            ("/bundle/issues/0029-collision.md", ISSUE_RECORD_0029),
+        ]);
+
+        let err = describe(
+            &store,
+            Path::new("/bundle"),
+            "0029",
+            "A concise description.",
+        )
+        .expect_err("an unqualified cross-type collision must be rejected");
+
+        assert!(err.contains("/bundle/adr/0029-collision.md"), "got: {err}");
+        assert!(
+            err.contains("/bundle/issues/0029-collision.md"),
+            "got: {err}"
+        );
+        let adr = store
+            .read(Path::new("/bundle/adr/0029-collision.md"))
+            .unwrap();
+        let issue = store
+            .read(Path::new("/bundle/issues/0029-collision.md"))
+            .unwrap();
+        assert_eq!(
+            adr, ADR_RECORD_0029,
+            "no file may be written on an ambiguous reference"
+        );
+        assert_eq!(
+            issue, ISSUE_RECORD_0029,
+            "no file may be written on an ambiguous reference"
+        );
     }
 
     #[test]
